# openapi/users.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/me": {
      "get": {
        "operationId": "User.me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/MeResponse"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Hồ sơ của tôi",
        "tags": [
          "Me"
        ]
      },
      "patch": {
        "operationId": "User.updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMe"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/MeResponse"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Sửa hồ sơ của tôi",
        "tags": [
          "Me"
        ]
      },
      "delete": {
        "operationId": "User.deleteMe",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Xoá tài khoản",
        "tags": [
          "Me"
        ]
      }
    },
    "/api/v1/admin/users": {
      "post": {
        "operationId": "UserAdmin.createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminUser"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Tạo tài khoản email + mật khẩu",
        "tags": [
          "Users"
        ],
        "description": "**Quyền cần có:** `user:create`"
      },
      "get": {
        "operationId": "UserAdmin.listUsers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminUser"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Danh sách user",
        "tags": [
          "Users"
        ],
        "description": "**Quyền cần có:** `user:read`"
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "UserAdmin.getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminUser"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Chi tiết một user",
        "tags": [
          "Users"
        ],
        "description": "**Quyền cần có:** `user:read`"
      }
    },
    "/api/v1/admin/users/{id}/status": {
      "patch": {
        "operationId": "UserAdmin.setUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminUser"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Khoá / mở khoá user",
        "tags": [
          "Users"
        ],
        "description": "**Quyền cần có:** `user:ban`"
      }
    },
...
  },
  "info": {
    "title": "C9 Map · User & Auth",
    "description": "Ngôn ngữ: header `Accept-Language: vi | en` (mặc định vi).",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "supabase": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "description": "Mã lỗi SCREAMING_SNAKE, client rẽ nhánh theo mã này"
          },
          "msg": {
            "type": "string",
            "description": "Câu đã dịch theo header Accept-Language (vi mặc định, en)"
          },
          "data": {
            "type": "string",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "description": "Gửi kèm khi báo lỗi để tra log"
              }
            },
            "required": [
              "requestId"
            ],
            "additionalProperties": {},
            "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
          }
        },
        "required": [
          "success",
          "code",
          "msg",
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "email": {
            "nullable": true,
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "displayName": {
            "type": "string"
          },
          "username": {
            "nullable": true,
            "type": "string"
          },
          "avatarUrl": {
            "nullable": true,
            "type": "string",
            "format": "uri"
          },
          "locale": {
            "type": "string"
          },
          "homeCityCode": {
            "nullable": true,
            "type": "string"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2,
              "maxLength": 30,
              "pattern": "^[a-z][a-z0-9_]*$"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "username",
          "avatarUrl",
          "locale",
          "homeCityCode",
          "phoneVerified",
          "status",
          "roles",
          "permissions"
        ],
        "additionalProperties": false
      },
      "UpdateMe": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "nullable": true,
            "type": "string",
            "maxLength": 500,
            "format": "uri"
          },
          "locale": {
            "type": "string",
            "enum": [
              "vi",
              "en"
            ]
          },
          "homeCityCode": {
            "nullable": true,
            "type": "string"
          }
        }
      },
      "CreateUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 72
          },
          "displayName": {
            "type": "string"
          },
          "roles": {
            "default": [
              "user"
            ],
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2,
              "maxLength": 30,
              "pattern": "^[a-z][a-z0-9_]*$"
            }
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "SetUserStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}